Move redux-toolkit example back into 1 file
