--- packages/demo-redux-toolkit/store/slices/subjectPage/model.ts
@@ -1,9 +1,0 @@
-interface SubjectPageData {
-    id: string;
-    name: string;
-    stateTimestamp: number;
-}
-
-export interface SubjectPageState {
-    data: SubjectPageData | null;
-}
